Server now generates maps that match clients.
Server posts permenant portals depending on the current map you are in. 
This allows us to add multiple realm portals and will allow for portals 
of cowardice in dungeons.

Fixed the way sprites are rotated so bullets draw correctly and almost 
match their hitboxes.

Made super profile animators update the underlying animator and make 
calls to draw from the underlying animator. This centralizes the 
drawing.

Added bounding rectangles to everything that uses profile animator.

Made it so entering portals is collision based. Allows you to enter the 
portal you are standing on rather than the first portal in the list.